Move utility to sam file
